Move query parameter test case from Android tests to unit tests
