+++ algoliasearch/src/test/java/com/algolia/search/saas/QueryTest.java
@@ -1,0 +1,77 @@
+/*
+ * Copyright (c) 2015 Algolia
+ * http://www.algolia.com/
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ */
+
+package com.algolia.search.saas;
+
+import org.junit.Test;
+
+import static junit.framework.Assert.assertEquals;
+import static junit.framework.Assert.assertTrue;
+
+public class QueryTest extends PowerMockTestCase {
+
+    @Override
+    public void setUp() throws Exception {
+        super.setUp();
+    }
+
+    @Override
+    public void tearDown() throws Exception {
+        super.tearDown();
+    }
+
+    @Test
+    public void testQueryExtraParameters() throws Exception {
+        // Accessor methods.
+        {
+            Query query = new Query();
+            query.set("extra", "value");
+            assertEquals("value", query.get("extra"));
+            assertEquals("extra=value", query.getQueryString());
+            query.set("abc", "def");
+            String url = query.getQueryString();
+            assertTrue(url.equals("abc=def&extra=value") || url.equals("extra=value&abc=def"));
+        }
+        // Percent escaping.
+        {
+            Query query = new Query();
+            query.set("%&=", "àéïôù");
+            assertEquals("%25%26%3D=%C3%A0%C3%A9%C3%AF%C3%B4%C3%B9", query.getQueryString());
+        }
+        // Overriding a typed query parameter.
+        {
+            Query query = new Query();
+            query.hitsPerPage = 100;
+            query.set("hitsPerPage", "666");
+            assertEquals("hitsPerPage=100&hitsPerPage=666", query.getQueryString());
+        }
+        // Deleting a parameter.
+        {
+            Query query = new Query();
+            query.set("abc", "ABC");
+            query.set("def", "DEF");
+            query.set("abc", null);
+            assertEquals("def=DEF", query.getQueryString());
+        }
+    }
+}
